requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==4.2.21
django-admin-interface==0.30.0
django-colorfield==0.14.0
djangorestframework==3.16.0
idna==3.10
pillow==11.2.1
python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.32.3
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.13.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.4.0
user-agents==2.2.0
